Use test code and data from git repo in CI
Before this, the test code and data from the portable source was used,
but we didn't want to add all the rtest reference files to the
portable source so this is a better solution. It also removes the need
to rebuild the portable soruce when testing local changes to the
tests during development.