remove unused include reference to lib/graph
The only possible lib/graph header imported by any of the tools is agxbuf.h, but
this is shadowed by the same named header in lib/cgraph. Having lib/graph in the
list of include directories here was a little dangerous as reordering would have
caused the tools to pick up lib/graph's copy. Furthermore as we are not
currently building lib/graph, this likely would have caused link failures.
Related to #1549.