deallocate memory before exiting mkdefs.c
The code in mkdef.c gets compiled and executed during Graphviz compilation. When
building with Address Sanitizer enabled, instrumentation is also applied to this
utility during compilation. Usually this is not a problem, but when configuring
ASan to exit with an error on leak detection it would cause the build step that
runs mkdefs to erroneously fail. We now explicitly clean up before exiting
mkdefs, making ASan happy in all configurations. Closes #1741.