fix: realign non-triangle remove_overlap() with its prototype
There are two definitions of remove_overlap(), one in use when triangulation or
GTS is enabled and the other in use when either of these is disabled. The latter
definition did not match its declaration. As a result, any attempt to use this
function would most likely have resulted in stack corruption. Fixes #1544.